Refactored: Datadog::Chunker --> Datadog::Core::Chunker
